requirements.txt
psycopg2-binary
pydantic
python-dotenv
SQLAlchemy
starlette
uvicorn
httpx
